+++ src/javaDB/MainController.java
@@ -1,0 +1,4 @@
+package javaDB;
+
+public class MainController {
+}
