+++ src/javaResources/CandidateDelete.java
@@ -1,5 +1,7 @@
 package javaResources;
 
+import javaDB.DB1;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -7,12 +9,11 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
 @WebServlet("/deleteCandidate")
-public class deleteCandidateServlet extends HttpServlet {
+public class CandidateDelete extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         String memberNumber = request.getParameter("memberNumber");
         System.out.println(memberNumber);
