+++ src/javaResources/CandidatePick.java
@@ -1,5 +1,8 @@
 package javaResources;
 
+import javaDB.DB1;
+import javaDB.PasserDAO;
+
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -8,13 +11,11 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
 import java.util.List;
 import java.util.Map;
 
 @WebServlet("/pick")
-public class MemberServlet extends HttpServlet {
+public class CandidatePick extends HttpServlet {
     private static final long serialVersionUID = 1L;
     private Connection connection;
     @Override
