+++ src/javaResources/CandidateStatus.java
@@ -1,4 +1,7 @@
 package javaResources;
+import javaDB.DB1;
+import javaDB.UserDAO2;
+
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,13 +10,11 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
 import java.util.List;
 import java.util.Map;
 //g
 @WebServlet("/candidateStatus")
-public class StatusServlet extends HttpServlet {
+public class CandidateStatus extends HttpServlet {
     private static final long serialVersionUID = 1L;
     private Connection connection;
     @Override
