+++ src/javaResources/CandidateUpdate.java
@@ -1,5 +1,7 @@
 package javaResources;
 
+import javaDB.DB1;
+
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -8,13 +10,12 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
 
 @WebServlet("/updateCandidate")
-public class updateCandidate extends HttpServlet {
+public class CandidateUpdate extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
